+++ crawl-ref/source/player.h
@@ -1238,7 +1238,7 @@ int spell_mp_freeze(spell_type which_spell);
 int weapon_sp_cost(const item_def* weapon, const item_def* ammo = nullptr);
 int player_tohit_modifier(const int old_tohit, int range = 1);
 int player_damage_modifier(int old_damage, bool silent = false, const int range = 1);
-int player_spellpower_modifier(int old_spellpower);
+double player_spellpower_modifier(double old_spellpower);
 int player_spellsuccess_modifier(int force);
 int player_attack_delay_modifier(int attk_delay);
 int player_stealth_modifier(int old_stealth);
